[Fix] 데미지값 자료형 수정
- 데미지, 체력은 int값이어야 하며, float로 잘못 표기된 부분을 수정했습니다.

+++ Assets/_Scripts/Character/Core/BaseCharacter.cs
@@ -3,14 +3,14 @@ using UnityEngine;
 public class BaseCharacter : MonoBehaviour
 {
     [Header("Character Status")]
-    [SerializeField] protected float damage;
-    [SerializeField] protected float currentHP;
-    [SerializeField] protected float maxHP;
+    [SerializeField] protected int damage;
+    [SerializeField] protected int currentHP;
+    [SerializeField] protected int maxHP;
 
 
-    public float Damage => damage;
-    public float CurrentHP => currentHP;
-    public float MaxHP => maxHP;
+    public int Damage => damage;
+    public int CurrentHP => currentHP;
+    public int MaxHP => maxHP;
 
     protected virtual void Awake()
     {
